midtrem/service_b/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
requests